# api-docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Information related to api exposed by teammacc system.",
    "version": "1.0",
    "title": "Stock API",
    "termsOfService": "https://www.teammacc.com/termsandconditions",
    "contact": {
      "name": "Team MACC",
      "url": "http://www.teammacc.com",
      "email": "support@teammacc.com"
    },
    "license": {
      "name": "License of API",
      "url": "https://www.teammacc.com/privacypolicy"
    }
  },
  "host": "localhost:8087",
  "tags": [
    {
      "name": "Stock",
      "description": "Stock Controller"
    }
  ],
  "paths": {
    "/stock/acrescentar/{idProduto}/{quantidade}": {
      "put": {
        "tags": [
          "Stock"
        ],
        "summary": "Acrescenta mais a quantidade informada para o estoque do produto informado",
        "operationId": "acrescentarUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "idProduto",
            "in": "path",
            "description": "Número do ID do produto que se deseja acrescentar estoque",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "quantidade",
            "in": "path",
            "description": "Quantidade do estoque que se deseja acrescentar",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Indica que o estoque do porduto especificado foi acrescido da quantidade especificada",
            "schema": {
              "$ref": "#/definitions/StockVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Nenhum registro encotrado para o ID especificado"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Foi gerada uma exceção no servidor"
          }
        },
        "security": [
          {
            "jwtToken": []
          }
        ]
      }
    },
    "/stock/atualizar/{idProduto}/{quantidade}": {
      "put": {
        "tags": [
          "Stock"
        ],
        "summary": "Salva um novo estoque no valor da quantidade informada para o produto informado",
        "operationId": "cadastrarUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "idProduto",
            "in": "path",
            "description": "Número do ID do produto que se deseja cadastrar estoque",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "quantidade",
            "in": "path",
            "description": "Quantidade do estoque que se deseja cadastrar",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Indica que o estoque do porduto especificado foi atualizado para a quantidade especificada",
            "schema": {
              "$ref": "#/definitions/StockVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Nenhum registro encotrado para o ID especificado"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Foi gerada uma exceção no servidor"
          }
        },
        "security": [
          {
            "jwtToken": []
          }
        ]
      }
    },
    "/stock/baixar/{idProduto}/{quantidade}": {
      "put": {
        "tags": [
          "Stock"
        ],
        "summary": "Reduz o estoque do produto informado na quantidade informada",
        "operationId": "baixarUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "idProduto",
            "in": "path",
            "description": "Número do ID do produto que se deseja baixar o estoque",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "quantidade",
            "in": "path",
            "description": "Quantidade do estoque que se deseja baixar",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Indica que o estoque do porduto especificado foi baixado da quantidade especificada",
            "schema": {
              "$ref": "#/definitions/StockVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Nenhum registro encotrado para o ID especificado"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Foi gerada uma exceção no servidor"
          }
        },
        "security": [
          {
            "jwtToken": []
          }
        ]
      }
    },
    "/stock/consultar/{idProduto}": {
      "get": {
        "tags": [
          "Stock"
        ],
        "summary": "Retorna a entidade especificada pelo ID do produto",
        "operationId": "findByIdProdutoUsingGET",
        "produces": [
          "application/json",
          "application/x-yaml",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "idProduto",
            "in": "path",
            "description": "Número do ID do produto que se deseja recuperar",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna a entidade especificada pelo idProduto",
            "schema": {
              "$ref": "#/definitions/StockVO"
            }
          },
          "400": {
            "description": "Nenhum registro encotrado para o ID especificado"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Foi gerada uma exceção no servidor"
          }
        },
        "security": [
          {
            "jwtToken": []
          }
        ]
      }
    },
    "/stock/verificar/{idProduto}/{quantidade}": {
      "get": {
        "tags": [
          "Stock"
        ],
        "summary": "Verifica se o produto informado possui em estoque, pelo menos, a quantidade informada",
        "operationId": "verificarUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "idProduto",
            "in": "path",
            "description": "Número do ID do produto que se deseja verificar estoque",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "quantidade",
            "in": "path",
            "description": "Quantidade do estoque que se deseja verificar a existência",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "true - Indica que o estoque do porduto especificado possui a quantidade especificada",
            "schema": {
              "type": "boolean"
            }
          },
          "400": {
            "description": "Nenhum registro encotrado para o ID especificado"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Foi gerada uma exceção no servidor"
          }
        },
        "security": [
          {
            "jwtToken": []
          }
        ]
      }
    },
    "/stock/{id}": {
      "get": {
        "tags": [
          "Stock"
        ],
        "summary": "Retorna a entidade especificada pelo ID",
        "operationId": "findByIdUsingGET",
        "produces": [
          "application/json",
          "application/x-yaml",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Número do ID do estoque que se deseja recuperar",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna a entidade especificada pelo ID",
            "schema": {
              "$ref": "#/definitions/StockVO"
            }
          },
          "400": {
            "description": "Nenhum registro encotrado para o ID especificado"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Foi gerada uma exceção no servidor"
          }
        },
        "security": [
          {
            "jwtToken": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "jwtToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "StockVO": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "idProduto": {
          "type": "integer",
          "format": "int64"
        },
        "quantidadeProduto": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "StockVO"
    }
  }
}